move database query tests into src/
