--- core/indexer/tests/get_blocks_query.rs
@@ -1,124 +1,0 @@
-use anyhow::Result;
-use indexer::{
-    database::{
-        queries::{get_blocks_paginated, insert_block},
-        types::BlockQuery,
-    },
-    test_utils::{new_mock_block_hash, new_test_db},
-};
-use indexer_types::BlockRow;
-
-#[tokio::test]
-async fn test_get_blocks_query() -> Result<()> {
-    let (_, writer, _temp_dir) = new_test_db().await?;
-    let conn = writer.connection();
-    insert_block(
-        &conn,
-        BlockRow::builder()
-            .height(100)
-            .hash(new_mock_block_hash(100))
-            .build(),
-    )
-    .await?;
-
-    insert_block(
-        &conn,
-        BlockRow::builder()
-            .height(101)
-            .hash(new_mock_block_hash(101))
-            .build(),
-    )
-    .await?;
-
-    insert_block(
-        &conn,
-        BlockRow::builder()
-            .height(102)
-            .hash(new_mock_block_hash(102))
-            .build(),
-    )
-    .await?;
-
-    let (blocks, meta) =
-        get_blocks_paginated(&conn, BlockQuery::builder().limit(1).build()).await?;
-
-    assert_eq!(blocks.len(), 1);
-    assert_eq!(blocks[0].height, 102);
-    assert!(meta.has_more);
-    assert_eq!(meta.next_cursor, Some(blocks[0].height));
-    assert_eq!(meta.total_count, 3);
-
-    let (blocks, meta) = get_blocks_paginated(
-        &conn,
-        BlockQuery::builder()
-            .maybe_cursor(meta.next_cursor)
-            .limit(1)
-            .build(),
-    )
-    .await?;
-
-    assert_eq!(blocks.len(), 1);
-    assert_eq!(blocks[0].height, 101);
-    assert!(meta.has_more);
-    assert_eq!(meta.next_cursor, Some(blocks[0].height));
-
-    let (blocks, meta) = get_blocks_paginated(
-        &conn,
-        BlockQuery::builder()
-            .maybe_cursor(meta.next_cursor)
-            .limit(1)
-            .build(),
-    )
-    .await?;
-
-    assert_eq!(blocks.len(), 1);
-    assert_eq!(blocks[0].height, 100);
-    assert!(!meta.has_more);
-    assert_eq!(meta.next_cursor, Some(blocks[0].height));
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_get_blocks_query_relevant() -> Result<()> {
-    let (_, writer, _temp_dir) = new_test_db().await?;
-    let conn = writer.connection();
-    insert_block(
-        &conn,
-        BlockRow::builder()
-            .height(100)
-            .hash(new_mock_block_hash(100))
-            .relevant(true)
-            .build(),
-    )
-    .await?;
-
-    insert_block(
-        &conn,
-        BlockRow::builder()
-            .height(101)
-            .hash(new_mock_block_hash(101))
-            .build(),
-    )
-    .await?;
-
-    let (blocks, meta) =
-        get_blocks_paginated(&conn, BlockQuery::builder().relevant(true).build()).await?;
-
-    assert_eq!(blocks.len(), 1);
-    assert_eq!(blocks[0].height, 100);
-    assert!(!meta.has_more);
-    assert_eq!(meta.next_cursor, Some(blocks[0].height));
-    assert_eq!(meta.total_count, 1);
-
-    let (blocks, meta) =
-        get_blocks_paginated(&conn, BlockQuery::builder().relevant(false).build()).await?;
-
-    assert_eq!(blocks.len(), 1);
-    assert_eq!(blocks[0].height, 101);
-    assert!(!meta.has_more);
-    assert_eq!(meta.next_cursor, Some(blocks[0].height));
-    assert_eq!(meta.total_count, 1);
-
-    Ok(())
-}
